Move timestamp_synopsis to its own header
