+++ libvast/CMakeLists.txt
@@ -104,6 +104,7 @@ set(libvast_sources
   src/table_slice.cpp
   src/table_slice_builder.cpp
   src/time.cpp
+  src/timestamp_synopsis.cpp
   src/to_events.cpp
   src/type.cpp
   src/uuid.cpp
