+++ libvast/vast/timestamp_synopsis.hpp
@@ -1,0 +1,28 @@
+/******************************************************************************
+ *                    _   _____   __________                                  *
+ *                   | | / / _ | / __/_  __/     Visibility                   *
+ *                   | |/ / __ |_\ \  / /          Across                     *
+ *                   |___/_/ |_/___/ /_/       Space and Time                 *
+ *                                                                            *
+ * This file is part of VAST. It is subject to the license terms in the       *
+ * LICENSE file found in the top-level directory of this distribution and at  *
+ * http://vast.io/license. No part of VAST, including this file, may be       *
+ * copied, modified, propagated, or distributed except according to the terms *
+ * contained in the LICENSE file.                                             *
+ ******************************************************************************/
+
+#pragma once
+
+#include "vast/min_max_synopsis.hpp"
+#include "vast/synopsis.hpp"
+
+namespace vast {
+
+class timestamp_synopsis final : public min_max_synopsis<timestamp> {
+public:
+  timestamp_synopsis(vast::type x);
+
+  bool equals(const synopsis& other) const noexcept override;
+};
+
+} // namespace vast
